pyprintf: Perform unsigned conversion correctly
The code requires conversting from a 32-bit C number to a python integer.
This conversion has to be aware of the signedness of the C number to perform
correct conversion.  In particular, the new code now matches the C behavior
for cases like:
pyprintf("%x", 0xFFFFFFF0);
The previous code interpreted this as "-16" and output a hex string like "-10"
The new code matches the C output of "FFFFFFF0"